fixed return type of dfs function from int8_t to int (was a typo)

+++ ALG 13-Can Go.cpp
@@ -17,7 +17,7 @@ bool valid(int i, int j)
     return true;
 }
 
-int8_t dfs(int si, int sj)
+int dfs(int si, int sj)
 {
     vis[si][sj] = true;
     if (grid[si][sj] == room)
